Cleaning up options handling and plumbing in deployments placeholder

+++ build.sh
@@ -5,11 +5,14 @@ cyan="\e[0;36m"
 
 TEMP=temp                             # Temp directory for build artifacts
 ISO=${TEMP}/iso                       # Build location for staging iso/boot files
-ROOT=${TEMP}/root                     # Root for installing packages for extraction
+ROOT=${TEMP}/root                     # Root mount point for layered filesystems
+BUILD=${TEMP}/build                   # Build location for packages extraction and builds
+LAYERS=${TEMP}/layers                 # Layered filesystems to include in the ISO
 GRUB=grub                             # Location to pull persisted Grub configuration files from
 BOOT_CFG="${ISO}/boot/grub/boot.cfg"  # Boot menu entries to read in
 
-check() {
+check()
+{
   if [ $? -ne 0 ]; then
     echo -e "${red}failed!${none}"
     exit 1
@@ -17,6 +20,29 @@ check() {
     echo -e "${cyan}success!${none}"
   fi
 }
+
+# Provide a failsafe for umounting mount points on exit
+RELEASED=0
+release()
+{
+  if [ $RELEASED -ne 1 ]; then
+    RELEASED=1
+    if mountpoint -q $BUILD; then
+      echo -en ":: Releasing mount point ${cyan}${BUILD}${none}..."
+      sudo umount -R $BUILD
+      check
+    fi
+  fi
+  exit
+}
+trap release EXIT
+trap release SIGINT
+
+# Ensure the current user has passwordless sudo access
+if ! sudo -l | grep -q "NOPASSWD: ALL"; then
+  echo -e ":: Passwordless sudo access is required see README.md..."
+  exit
+fi
 
 # Configure build environment
 # `coreutils`             provides basic linux tooling
@@ -30,36 +56,37 @@ check() {
 # `gptfdisk`              used by the installer to prepare target media for install
 # `linux`                 need to load the kernel to satisfy GRUB
 # `intel-ucode`           standard practice to load the intel-ucode
-config_build_env() {
-  echo -en ":: Configuring build environment..."
-  mkdir -p $ISO/boot/grub
-  check
-  if [ ! -d $ROOT ]; then
-    echo -e ":: Installing build environment dependencies..."
-    sudo mkdir -p $ROOT
-    sudo pacstrap -c -G -M $ROOT coreutils pacman grub sed linux intel-ucode memtest86+ mkinitcpio \
+build_env()
+{
+  if [ ! -d $BUILD ]; then
+    echo -en ":: Configuring build environment..."
+    sudo mkdir -p $BUILD
+    sudo pacstrap -c -G -M $BUILD coreutils pacman grub sed linux intel-ucode memtest86+ mkinitcpio \
       mkinitcpio-vt-colors dosfstools rsync gptfdisk
   fi
+}
 
-  # Install kernel, intel ucode patch and memtest
+# Configure grub theme and build supporting BIOS and EFI boot images required to make
+# the ISO bootable as a CD or USB stick on BIOS and UEFI systems with the same presentation.
+build_multiboot()
+{
+  echo -e ":: Building multiboot components..."
+  mkdir -p $ISO/boot/grub/themes
+
   echo -en ":: Copying kernel, intel ucode patch and memtest to ${ISO}/boot..."
-  cp $ROOT/boot/intel-ucode.img $ISO/boot
-  cp $ROOT/boot/vmlinuz-linux $ISO/boot
-  cp $ROOT/boot/memtest86+/memtest.bin $ISO/boot/memtest
+  cp $BUILD/boot/intel-ucode.img $ISO/boot
+  cp $BUILD/boot/vmlinuz-linux $ISO/boot
+  cp $BUILD/boot/memtest86+/memtest.bin $ISO/boot/memtest
   check
 
-  # Install GRUB config and theme
-  mkdir -p $ISO/boot/grub/themes
   echo -en ":: Copying GRUB config and theme to ${ISO}/boot/grub ..."
   cp $GRUB/grub.cfg $ISO/boot/grub
   cp $GRUB/boot.cfg $ISO/boot/grub
   cp $GRUB/loopback.cfg $ISO/boot/grub
   cp -r $GRUB/themes $ISO/boot/grub
-  cp $ROOT/usr/share/grub/unicode.pf2 $ISO/boot/grub
+  cp $BUILD/usr/share/grub/unicode.pf2 $ISO/boot/grub
   check
 
-  # Create the GRUB menu items
-  # Pass through options as kernel paramaters
   echo -e ":: Creating ${cyan}${ISO}/boot/grub/boot.cfg${none} ..."
   echo -e "menuentry --class=deployment 'Start cyberlinux live' {" >> ${BOOT_CFG}
   echo -e "  cat /boot/grub/themes/cyberlinux/splash" >> ${BOOT_CFG}
@@ -67,43 +94,36 @@ config_build_env() {
   echo -e "  linux	/boot/vmlinuz-linux kernel=linux layers=shell,lite" >> ${BOOT_CFG}
   echo -e "  initrd	/boot/intel-ucode.img /boot/installer" >> ${BOOT_CFG}
   echo -e "}" >> ${BOOT_CFG}
-}
 
-# Install BIOS boot files
-create_bios_boot_images() {
-  echo -en ":: Creating core BIOS $ROOT/bios.img..."
-  cp -r $ROOT/usr/lib/grub/i386-pc $ISO/boot/grub
+  echo -en ":: Creating core BIOS $BUILD/bios.img..."
+  cp -r $BUILD/usr/lib/grub/i386-pc $ISO/boot/grub
   rm -f $ISO/boot/grub/i386-pc/*.img
   # We need to create our bios.img that contains just enough code to find the grub configuration and
   # grub modules in /boot/grub/i386-pc directory we'll stage in the next step
   # -O i386-pc                        Format of the image to generate
   # -p /boot/grub                     Directory to find grub once booted
-  # -d $ROOT/usr/lib/grub/i386-pc     Use resources from this location when building the boot image
-  # -o $ROOT/bios.img                 Output destination
-  grub-mkimage -O i386-pc -p /boot/grub -d $ROOT/usr/lib/grub/i386-pc -o $TEMP/bios.img  \
+  # -d $BUILD/usr/lib/grub/i386-pc    Use resources from this location when building the boot image
+  # -o $BUILD/bios.img                Output destination
+  grub-mkimage -O i386-pc -p /boot/grub -d $BUILD/usr/lib/grub/i386-pc -o $TEMP/bios.img  \
     biosdisk disk part_msdos part_gpt linux linux16 loopback normal configfile test search search_fs_uuid \
     search_fs_file true iso9660 search_label gfxterm gfxmenu gfxterm_menu ext2 ntfs cat echo ls memdisk tar
   check
   echo -en ":: Concatenate cdboot.img to bios.img to create the CD-ROM bootable Eltorito $TEMP/eltorito.img..."
-  cat $ROOT/usr/lib/grub/i386-pc/cdboot.img $TEMP/bios.img > $ISO/boot/grub/i386-pc/eltorito.img
+  cat $BUILD/usr/lib/grub/i386-pc/cdboot.img $TEMP/bios.img > $ISO/boot/grub/i386-pc/eltorito.img
   check
   echo -en ":: Concatenate boot.img to bios.img to create isohybrid $TEMP/isohybrid.img..."
-  cat $ROOT/usr/lib/grub/i386-pc/boot.img $TEMP/bios.img > $ISO/boot/grub/i386-pc/isohybrid.img
+  cat $BUILD/usr/lib/grub/i386-pc/boot.img $TEMP/bios.img > $ISO/boot/grub/i386-pc/isohybrid.img
   check
-}
 
-# Install UEFI boot files
-# /efi/boot/bootx64.efi is a well known boot file location for compatibility
-create_efi_boot_images() {
   echo -en ":: Creating UEFI boot files..."
   mkdir -p $ISO/efi/boot
-  cp -r $ROOT/usr/lib/grub/x86_64-efi $ISO/boot/grub
+  cp -r $BUILD/usr/lib/grub/x86_64-efi $ISO/boot/grub
   rm -f $ISO/grub/x86_64-efi/*.img
   # -O x86_64-efi                     Format of the image to generate
   # -p /boot/grub                     Directory to find grub once booted
-  # -d $ROOT/usr/lib/grub/x86_64-efi  Use resources from this location when building the boot image
+  # -d $BUILD/usr/lib/grub/x86_64-efi  Use resources from this location when building the boot image
   # -o $ISO/efi/boot/bootx64.efi      Output destination, using wellknown compatibility location
-  grub-mkimage -O x86_64-efi -p /boot/grub -d $ROOT/usr/lib/grub/x86_64-efi -o $ISO/efi/boot/bootx64.efi \
+  grub-mkimage -O x86_64-efi -p /boot/grub -d $BUILD/usr/lib/grub/x86_64-efi -o $ISO/efi/boot/bootx64.efi \
     disk part_msdos part_gpt linux linux16 loopback normal configfile test search search_fs_uuid \
     search_fs_file true iso9660 search_label efi_uga efi_gop gfxterm gfxmenu gfxterm_menu ext2 \
     ntfs cat echo ls memdisk tar
@@ -113,13 +133,16 @@ create_efi_boot_images() {
 # Build the initramfs based installer
 build_installer() {
   echo -en ":: Build the initramfs based installer..."
-  sudo cp installer/installer $ROOT/usr/lib/initcpio/hooks
-  sudo cp installer/installer.conf $ROOT/usr/lib/initcpio/install/installer
-  sudo cp installer/mkinitcpio.conf $ROOT/etc
-  sudo mount --bind $ROOT $ROOT
-  sudo arch-chroot $ROOT mkinitcpio -g /root/installer
-  sudo cp $ROOT/root/installer $ISO/boot
-  sudo umount $ROOT
+  mkdir -p $ISO/boot
+  sudo cp installer/installer $BUILD/usr/lib/initcpio/hooks
+  sudo cp installer/installer.conf $BUILD/usr/lib/initcpio/install/installer
+  sudo cp installer/mkinitcpio.conf $BUILD/etc
+
+  # Mount as bind mount to satisfy arch-chroot requirement
+  # umount is handled by the release function on exit
+  sudo mount --bind $BUILD $BUILD
+  sudo arch-chroot $BUILD mkinitcpio -g /root/installer
+  sudo cp $BUILD/root/installer $ISO/boot
   check
 }
 
@@ -172,47 +195,67 @@ build_iso() {
     -o boot.iso $ISO
 }
 
+# Build deployments
+build_deployments() 
+{
+  echo -e ":: Building deployments ${cyan}${1}${none}..."
+}
+
 # Main entry point
 # -------------------------------------------------------------------------------------------------
-build() {
-  config_build_env
-  create_bios_boot_images
-  create_efi_boot_images
+header() {
+  echo -e "${cyan}CYBERLINUX${none} builder automation for a multiboot installer ISO"
 }
 usage() {
-  echo -e "${cyan}CYBERLINUX${none} initramfs based installer automation"
-  echo -e "Usage: ${cyan}./`basename $0`${none} [options]"
+  header
+  echo -e "Usage: ${cyan}./$(basename $0)${none} [options]"
   echo -e "Options:"
-  echo -e "-c  Clean the build artifacts before building"
-  echo -e "-f  Force boot images to be rebuilt"
-  echo -e "-i  Build the initramfs"
-  echo -e "-h  Display usage help"
+  echo -e "-a               Build all buildable options"
+  echo -e "-d DEPLOYMENTS   Build deployments, comma delimited (all|shell|lite)"
+  echo -e "-i               Build the initramfs installer"
+  echo -e "-m               Build the grub multiboot environment"
+  echo -e "-c               Clean the build artifacts before building"
+  echo -e "-h               Display usage help\n"
+  echo -e "Examples:"
+  echo -e "Build everything: ./$(basename $0) -a"
   exit 1
 }
-while getopts ":cfih" opt; do
+while getopts ":ad:imch" opt; do
   case $opt in
-    c)
-      echo "Cleaning build artifacts before building"
-      sudo rm -rf $TEMP 
-      ;;
-    f)
-      build
-      ;;
-    i)
-      [ ! -d $TEMP ] && build
-      build_installer
-      ;;
-    h)
-      usage
-      ;;
-    \?)
-      echo -e "Invalid option: ${red}-$OPTARG${none}\n"
-      usage
-      ;;
+    c) CLEAN=1;;
+    a) ALL=1;;
+    i) INSTALLER=1;;
+    d) DEPLOYMENTS=$OPTARG;;
+    m) MULTIBOOT=1;;
+    h) usage;;
+    \?) echo -e "Invalid option: ${red}-${OPTARG}${none}\n"; usage;;
+    :) echo -e "Option ${red}-${OPTARG}${none} requires an argument\n"; usage;;
   esac
 done
+[ $(($OPTIND -1)) -eq 0 ] && usage
 
-[ ! -d $TEMP ] && build
-build_iso
+# Execute build
+header
+if [ ! -z ${CLEAN+x} ]; then
+  echo -e ":: Cleaning build artifacts before building"
+  sudo rm -rf $TEMP
+fi
+mkdir -p $TEMP
+
+# Always build the build environment if any build option is chosen
+if [ ! -z ${ALL+x} ] || [ ! -z ${MULTIBOOT+x} ] || [ ! -z ${INSTALLER+x} ] || [ ! -z ${DEPLOYMENTS+x} ]; then
+  build_env
+fi
+
+# Needs to happen before the multiboot as deployments will be boot entries
+if [ ! -z ${ALL+x} ] || [ ! -z ${DEPLOYMENTS+x} ]; then
+  build_deployments $DEPLOYMENTS
+fi
+if [ ! -z ${ALL+x} ] || [ ! -z ${MULTIBOOT+x} ]; then
+  build_multiboot
+fi
+if [ ! -z ${ALL+x} ] || [ ! -z ${INSTALLER+x} ]; then
+  build_installer
+fi
 
 # vim: ft=sh:ts=2:sw=2:sts=2
